fix: add item directly to stack and improve srp adherence
