+++ src/design/forms/verificationForm/verificationBoxes/renderStepTwoVerification/index.tsx
@@ -1,33 +1,10 @@
 import { Grid, Stack, Tag } from "@inubekit/inubekit";
 import { BoxAttribute } from "@design/feedback/boxAttribute";
 import { basic } from "@design/tokens";
-
-import { ComponentAppearance } from "@ptypes/aparences.types";
-import { StyldTagContainer } from "./styles";
 import { labels } from "@config/verificationTitles";
 import { IOptionInitialiceEntry } from "@ptypes/positions/assisted/IOptionInitialiceEntry";
-import { IGeneralInformationEntry } from "@ptypes/positions/assisted/IGeneralInformationEntry";
-
-const renderPersonalInfoVerification = (
-  values: IGeneralInformationEntry,
-  isMobile: boolean
-) => (
-  <>
-    <Grid
-      templateColumns={isMobile ? "1fr" : "1fr "}
-      autoRows="1fr"
-      width="100%"
-    >
-      <BoxAttribute label={labels.namePosition} value={values.namePosition} />
-    </Grid>
-    <Stack width="100%" direction="column" gap={basic.spacing.s100}>
-      <BoxAttribute
-        label={labels.descriptionPosition}
-        value={values.descriptionPosition}
-      />
-    </Stack>
-  </>
-);
+import { ComponentAppearance } from "@ptypes/aparences.types";
+import { StyldTagContainer } from "../styles";
 
 const renderStepTwoVerification = (
   values: IOptionInitialiceEntry[],
@@ -65,4 +42,4 @@ const renderStepTwoVerification = (
   );
 };
 
-export { renderPersonalInfoVerification, renderStepTwoVerification };
+export default renderStepTwoVerification;
