+++ src/design/modals/detailsRequestsInProgressModal/styles.ts
@@ -18,18 +18,6 @@ const StyledModal = styled.div<IStyledModal>`
   height: 100%;
 `;
 
-const StyledModalTraceability = styled.div<IStyledModal>`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: ${basic.spacing.s150};
-  background-color: ${inube.palette.neutral.N0};
-  border: 1px solid ${inube.palette.neutral.N30};
-  border-radius: 8px;
-  width: ${({ $smallScreen }) => ($smallScreen ? "auto" : "400px")};
-  padding: ${basic.spacing.s300};
-`;
-
 const StyledModalConatiner = styled.div`
   display: flex;
   flex-direction: column;
@@ -49,14 +37,6 @@ const StyledContainerButton = styled.div`
   }
 `;
 
-const StyledContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid ${inube.palette.neutral.N30};
-  gap: ${basic.spacing.s200};
-  width: 100%;
-`;
-
 const StyledContainerData = styled.div`
   display: flex;
   flex-direction: column;
@@ -69,34 +49,9 @@ const StyledContainerData = styled.div`
   background-color: ${inube.palette.neutral.N10};
 `;
 
-const StyledContainerDataTraceability = styled.div`
-  display: flex;
-  flex-direction: column;
-  border: 1px solid ${inube.palette.neutral.N30};
-  padding: ${basic.spacing.s150};
-  gap: ${basic.spacing.s4};
-  border-radius: 8px;
-  border: 1px solid ${inube.palette.neutral.N10};
-  background-color: ${inube.palette.neutral.N10};
-  width: 100%;
-  box-sizing: border-box;
-`;
-
-const StyledContainerDataTraceabilitydos = styled.div`
-  width: 107px;
-  border: 1px solid ${inube.palette.neutral.N30};
-  padding: ${basic.spacing.s025} ${basic.spacing.s10};
-  background-color: ${inube.palette.neutral.N30};
-  border-radius: 4px;
-`;
-
 export {
   StyledModal,
   StyledContainerButton,
-  StyledModalTraceability,
   StyledModalConatiner,
-  StyledContainer,
   StyledContainerData,
-  StyledContainerDataTraceability,
-  StyledContainerDataTraceabilitydos,
 };
